# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tree Story Project",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/discovery/feed": {
      "get": {
        "tags": [
          "Discovery"
        ],
        "summary": "最新动态 (瀑布流)",
        "description": "获取全站最新发布的节点。",
        "operationId": "getLatestFeed",
        "parameters": [
          {
            "name": "book_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "[可选] 只看某个活动/书本的动态",
              "title": "Book Id"
            },
            "description": "[可选] 只看某个活动/书本的动态"
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StoryNodeListItem"
                  },
                  "title": "Response Getlatestfeed"
                }
              }
            }
          },
          "422": {
            "description": "参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/discovery/trending": {
      "get": {
        "tags": [
          "Discovery"
        ],
        "summary": "热门分支榜",
        "description": "获取 \"近期最热\" 的节点。\n算法：在最近 N 天内发布的节点中，按 likes_count 倒序排列。",
        "operationId": "getTrendingNodes",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 30,
              "minimum": 1,
              "description": "统计最近几天的热度 (1-30天)",
              "default": 7,
              "title": "Days"
            },
            "description": "统计最近几天的热度 (1-30天)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StoryNodeListItem"
                  },
                  "title": "Response Gettrendingnodes"
                }
              }
            }
          },
          "422": {
            "description": "参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/discovery/search": {
      "get": {
        "tags": [
          "Discovery"
        ],
        "summary": "关键词搜索",
        "description": "简单的模糊搜索 (LIKE %q%)。",
        "operationId": "searchNodes",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 50,
              "description": "搜索关键词",
              "title": "Q"
            },
            "description": "搜索关键词"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StoryNodeListItem"
                  },
                  "title": "Response Searchnodes"
                }
              }
            }
          },
          "422": {
            "description": "参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AuthorInfo": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "avatar": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username"
        ],
        "title": "AuthorInfo"
      },
...
      "NodeStatus": {
        "type": "string",
        "enum": [
          "pending",
          "published",
          "locked",
          "rejected"
        ],
        "title": "NodeStatus"
      },
...
      "StoryNodeListItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "book_id": {
            "type": "integer",
            "title": "Book Id"
          },
          "author": {
            "$ref": "#/components/schemas/AuthorInfo"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "branch_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Branch Name"
          },
          "status": {
            "$ref": "#/components/schemas/NodeStatus"
          },
          "depth": {
            "type": "integer",
            "title": "Depth"
          },
          "likes_count": {
            "type": "integer",
            "title": "Likes Count"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "book_id",
          "author",
          "status",
          "depth",
          "likes_count",
          "created_at"
        ],
        "title": "StoryNodeListItem",
        "description": "用于列表展示（feed/search/user-nodes），不含 children，不含 content。"
      },
...
      "ValidationErrorItem": {
        "properties": {
          "loc": {
            "items": {

            },
            "type": "array",
            "title": "Loc"
          },
          "msg": {
            "type": "string",
            "title": "Msg"
          },
          "type": {
            "type": "string",
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationErrorItem"
      },
      "ValidationErrorResponse": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationErrorItem"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ValidationErrorResponse"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {

            },
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}